use macro to gen runtime params
