+++ runtime/common/src/lib.rs
@@ -23,6 +23,7 @@ mod impl_on_charge_evm_transaction;
 mod impl_self_contained_call;
 mod impl_xcm_evm_runner;
 pub mod migrations;
+pub mod runtime_params;
 pub mod timestamp;
 pub mod types;
 pub mod weights;
